call the new tipwell compact so more clear
