+++ projects/novo-examples/src/components/tip-well/compact-tip-well/compact-tip-well-example.css
