+++ projects/novo-examples/src/components/tip-well/compact-tip-well/index.ts
@@ -1,0 +1,1 @@
+export * from './compact-tip-well-example';
