--- projects/novo-examples/src/components/tip-well/icon-button-tip-well/index.ts
@@ -1,1 +1,0 @@
-export * from './icon-button-tip-well-example';
