+++ projects/novo-examples/src/components/tip-well/index.ts
@@ -2,4 +2,4 @@ export * from './basic-tip-well';
 export * from './buttonless-tip-well';
 export * from './html-tip-well';
 export * from './icon-tip-well';
-export * from './icon-button-tip-well';
+export * from './compact-tip-well';
